split frontend/binaryninja/handlers/conditions.py into two files
