+++ decompiler/frontend/binaryninja/handlers/conditions.py
@@ -3,11 +3,11 @@ from functools import partial
 
 from binaryninja import mediumlevelil
 from decompiler.frontend.lifter import Handler
-from decompiler.structures.pseudo import Branch, Condition, Constant, IndirectBranch, OperationType, Return
+from decompiler.structures.pseudo import Condition, OperationType
 
 
 class ConditionHandler(Handler):
-    """Handler for mlil conditions and branches,"""
+    """Handler for mlil conditions."""
 
     def register(self):
         """Register the handler functions at the parent lifter."""
@@ -31,12 +31,6 @@ class ConditionHandler(Handler):
                 mediumlevelil.MediumLevelILFcmpLt: partial(self.lift_condition, operation=OperationType.less),
                 mediumlevelil.MediumLevelILFcmpO: partial(self.lift_condition, operation=OperationType.equal),
                 mediumlevelil.MediumLevelILFcmpUo: partial(self.lift_condition, operation=OperationType.equal),
-                mediumlevelil.MediumLevelILRet: self.lift_return,
-                mediumlevelil.MediumLevelILIf: self.lift_branch,
-                mediumlevelil.MediumLevelILJump: lambda x: None,
-                mediumlevelil.MediumLevelILJumpTo: self.lift_branch_indirect,
-                mediumlevelil.MediumLevelILGoto: lambda x: None,
-                mediumlevelil.MediumLevelILNoret: lambda x: None,
             }
         )
 
@@ -45,18 +39,3 @@ class ConditionHandler(Handler):
         return Condition(
             operation, [self._lifter.lift(condition.left, parent=condition), self._lifter.lift(condition.right, parent=condition)]
         )
-
-    def lift_branch(self, branch: mediumlevelil.MediumLevelILIf, **kwargs) -> Branch:
-        """Lift a branch instruction by lifting its condition."""
-        condition = self._lifter.lift(branch.condition, parent=branch)
-        if not isinstance(condition, Condition):
-            condition = Condition(OperationType.not_equal, [condition, Constant(0, condition.type.copy())])
-        return Branch(condition)
-
-    def lift_branch_indirect(self, branch: mediumlevelil.MediumLevelILJumpTo, **kwargs) -> IndirectBranch:
-        """Lift a non-trivial jump instruction."""
-        return IndirectBranch(self._lifter.lift(branch.dest, parent=branch))
-
-    def lift_return(self, ret_op: mediumlevelil.MediumLevelILRet, **kwargs) -> Return:
-        """Lift a return instruction."""
-        return Return([self._lifter.lift(return_value, parent=ret_op) for return_value in ret_op.src])
